fix : 사건 pdf 관리하는 컨트롤러 별도로 분리 좀 더 명확한 구조를 만들기 위함
