+++ src/main/java/com/avg/lawsuitmanagement/lawsuit/service/LawsuitService.java
@@ -20,17 +20,10 @@ import com.avg.lawsuitmanagement.lawsuit.dto.BasicLawsuitDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.BasicUserDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.GetClientLawsuitListDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.GetEmployeeLawsuitListDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.IdNameDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitBasicDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitBasicRawDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitCountDto;
 import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitPrintAdviceDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitPrintAdvicePreDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitPrintExpenseDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitPrintLawsuitDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitPrintRawDto;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitPrintResponseDto;
 import com.avg.lawsuitmanagement.lawsuit.repository.LawsuitMapperRepository;
 import com.avg.lawsuitmanagement.lawsuit.repository.param.InsertLawsuitClientMemberIdParam;
 import com.avg.lawsuitmanagement.lawsuit.repository.param.InsertLawsuitParam;
@@ -42,7 +35,6 @@ import com.avg.lawsuitmanagement.lawsuit.type.LawsuitStatus;
 import com.avg.lawsuitmanagement.member.dto.MemberDto;
 import com.avg.lawsuitmanagement.member.dto.MemberDtoNonPass;
 import com.avg.lawsuitmanagement.member.repository.MemberMapperRepository;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -237,13 +229,11 @@ public class LawsuitService {
             employeeMap.put(employeeId, employee);
         }
 
-        LawsuitBasicDto result = LawsuitBasicDto.builder()
+        return LawsuitBasicDto.builder()
             .lawsuit(lawsuit)
             .employees(employeeMap.values().stream().toList())
             .clients(clientMap.values().stream().toList())
             .build();
-
-        return result;
     }
 
     public void updateStatus(Long id, LawsuitStatus status) {
@@ -251,128 +241,6 @@ public class LawsuitService {
             .id(id)
             .status(status.toString())
             .build());
-    }
-
-    public LawsuitPrintResponseDto getPrintInfo(Long lawsuitId) {
-        List<LawsuitPrintRawDto> raws = lawsuitMapperRepository.selectPrintInfo(
-            lawsuitId);
-
-        if (raws.isEmpty()) {
-            throw new RuntimeException("");
-        }
-        LawsuitPrintRawDto lawsuit = raws.get(0);
-
-        Map<Long, String> clientNameMap = new HashMap<>();
-        Map<Long, String> memberNameMap = new HashMap<>();
-        Map<Long, LawsuitPrintAdvicePreDto> adviceMap = new HashMap<>();
-        Map<Long, LawsuitPrintExpenseDto> expenseMap = new HashMap<>();
-        for (LawsuitPrintRawDto raw : raws) {
-            Long clientId = raw.getClientId();
-            if (clientId != null) {
-                if (!clientNameMap.containsKey(clientId)) {
-                    clientNameMap.put(clientId, raw.getClientName());
-                }
-            }
-
-            Long memberId = raw.getMemberId();
-            if (memberId != null) {
-                if (!memberNameMap.containsKey(memberId)) {
-                    memberNameMap.put(memberId, raw.getMemberName());
-                }
-            }
-
-            Long adviceId = raw.getAdviceId();
-            if (adviceId != null) {
-                if (!adviceMap.containsKey(adviceId)) {
-                    List<IdNameDto> memberIdNames = new ArrayList<>();
-                    if (raw.getMemberName() != null) {
-                        memberIdNames.add(IdNameDto.builder()
-                            .id(raw.getMemberId())
-                            .name(raw.getMemberName())
-                            .build());
-                    }
-                    List<IdNameDto> clientIdNames = new ArrayList<>();
-                    if (raw.getClientName() != null) {
-                        clientIdNames.add(IdNameDto.builder()
-                            .id(raw.getClientId())
-                            .name(raw.getClientName())
-                            .build());
-                    }
-                    adviceMap.put(adviceId, LawsuitPrintAdvicePreDto.builder()
-                        .id(raw.getAdviceId())
-                        .title(raw.getAdviceTitle())
-                        .contents(raw.getAdviceContents())
-                        .date(raw.getAdviceDate())
-                        .memberIdNames(memberIdNames)
-                        .clientIdNames(clientIdNames)
-                        .build());
-                } else {
-                    LawsuitPrintAdvicePreDto lawsuitPrintAdvicePreDto = adviceMap.get(adviceId);
-
-                    List<IdNameDto> memberIdNames = lawsuitPrintAdvicePreDto.getMemberIdNames();
-                    IdNameDto memberIdName = IdNameDto.builder()
-                        .id(raw.getMemberId())
-                        .name(raw.getMemberName())
-                        .build();
-                    if (raw.getMemberName() != null && memberIdNames.stream()
-                        .noneMatch(it -> it.getId()
-                            .equals(memberIdName.getId()))) {
-                        memberIdNames.add(memberIdName);
-                    }
-
-                    List<IdNameDto> clientIdNames = lawsuitPrintAdvicePreDto.getClientIdNames();
-                    IdNameDto clientIdName = IdNameDto.builder()
-                        .id(raw.getClientId())
-                        .name(raw.getClientName())
-                        .build();
-                    if (raw.getClientName() != null && clientIdNames.stream()
-                        .noneMatch(it -> it.getId()
-                            .equals(clientIdName.getId()))) {
-                        clientIdNames.add(clientIdName);
-                    }
-                }
-            }
-
-            Long expenseId = raw.getExpenseId();
-            if (expenseId != null) {
-                if (!expenseMap.containsKey(expenseId)) {
-                    expenseMap.put(expenseId, LawsuitPrintExpenseDto.builder()
-                        .id(raw.getExpenseId())
-                        .contents(raw.getExpenseContents())
-                        .amount(raw.getExpenseAmount())
-                        .date(raw.getAdviceDate())
-                        .build());
-                }
-            }
-        }
-
-        LawsuitPrintResponseDto result = LawsuitPrintResponseDto.builder()
-            .lawsuit(LawsuitPrintLawsuitDto.builder()
-                .id(lawsuit.getLawsuitId())
-                .name(lawsuit.getLawsuitName())
-                .num(lawsuit.getLawsuitNum())
-                .type(lawsuit.getLawsuitType())
-                .court(lawsuit.getCourtName())
-                .commissionFee(lawsuit.getLawsuitCommissionFee())
-                .contingentFee(lawsuit.getLawsuitContingentFee())
-                .judgementResult(lawsuit.getLawsuitJudgementResult())
-                .judgementDate(lawsuit.getLawsuitJudgementDate())
-                .clients(clientNameMap.values().stream().toList())
-                .members(memberNameMap.values().stream().toList())
-                .build())
-            .advices(adviceMap.values().stream().map(it -> LawsuitPrintAdviceDto.builder()
-                    .id(it.getId())
-                    .title(it.getTitle())
-                    .contents(it.getContents())
-                    .date(it.getDate())
-                    .memberNames(it.getMemberIdNames().stream().map(IdNameDto::getName).toList())
-                    .clientNames(it.getClientIdNames().stream().map(IdNameDto::getName).toList())
-                    .build())
-                .toList())
-            .expenses(expenseMap.values().stream().toList())
-            .build();
-
-        return result;
     }
 
     private SelectClientLawsuitListParam getParam(GetClientLawsuitForm form, long clientId) {
